requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.15
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.0.4
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
whitenoise==6.2.0
zipp==3.8.1
